+++ src/components/common/calendar/Calendar.tsx
@@ -1,7 +1,5 @@
-import React, { useEffect, useState } from "react";
-import dayjs from "dayjs";
-
 import { ArrowIcon } from "assets";
+import useCalendar from "./hooks/useCalendar";
 import * as S from "./Calendar.styled";
 
 interface CalendarProps {
@@ -9,56 +7,17 @@ interface CalendarProps {
 }
 
 const Calendar = ({ dates }: CalendarProps) => {
-  const [currentMonth, setCurrentMonth] = useState(dayjs());
-  const [period, setPeriod] = useState("");
-  const [disabled, setDisabled] = useState<string[]>([]);
+  const {
+    currentMonth,
+    monthsData,
+    disabled,
+    calendarList,
+    handleMoveToPrevMonth,
+    handleMoveToNextMonth,
+    handleDisabledCheck,
+  } = useCalendar(dates);
 
   const dayWeek = ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"] as const;
-
-  const daysInMonth = currentMonth.daysInMonth();
-  const firstDayOfMonth = dayjs(currentMonth).startOf("month").locale("ko");
-  const currentDate = dayjs(dayjs()).add(+period, "day");
-  const existDates = Array.from({ length: daysInMonth }, (_, index) =>
-    dayjs(firstDayOfMonth).add(index, "day")
-  );
-  const emptyDates = new Array(firstDayOfMonth.day()).fill(null);
-  const calendarList = [...emptyDates, ...existDates];
-
-  const monthsData: string[] = [];
-  let currentDay = dayjs().startOf("day");
-
-  while (currentDay.isSameOrBefore(currentDate.startOf("day"))) {
-    monthsData.push(currentDay.format("YYYY-MM-DD"));
-    currentDay = currentDay.add(1, "day");
-  }
-
-  const handleMoveToPrevMonth = (): void => {
-    setCurrentMonth(dayjs(currentMonth).subtract(1, "month"));
-  };
-  const handleMoveToNextMonth = (): void => {
-    setCurrentMonth(dayjs(currentMonth).add(1, "month"));
-  };
-
-  const handleDisabledCheck = (date: dayjs.Dayjs) => (): void => {
-    if (!date) return;
-
-    const formattedDate = date.format("YYYY-MM-DD");
-    const isAlreadyDisabled = disabled.includes(formattedDate);
-
-    if (!isAlreadyDisabled) {
-      if (!monthsData.includes(formattedDate)) return;
-
-      setDisabled([...disabled, formattedDate]);
-    } else {
-      const currentDate = disabled.filter((item) => item !== formattedDate);
-
-      setDisabled(currentDate);
-    }
-  };
-
-  useEffect(() => {
-    setPeriod(dates);
-  }, [dates]);
 
   return (
     <S.Calendar>
